Move section to ui folder
