+++ src/lib/components/account/Import.svelte
@@ -2,7 +2,7 @@
 	import { db } from '$lib/db';
 	import { t } from '$lib/translations';
 	import { RefreshCw } from 'lucide-svelte';
-	import * as Section from './section';
+	import * as Section from '../ui/section';
 
 	let error = $state('');
 	let uploading = $state(false);
