+++ src/lib/components/ui/section/Description.svelte
